requirements.txt
# ============================================================
# Vertex Football Scraper 2 - Full Dependencies
# ============================================================

# --- Networking & HTTP (Async & Cloudflare Bypass) ---
aiohttp>=3.9.0             # Async HTTP for Understat
curl_cffi>=0.6.0           # Cloudflare bypass for SofaScore/FBref
nodriver>=0.40.0           # Undetected Chrome for browser-based automation
requests>=2.31.0           # Standard synchronous HTTP

# --- HTML Parsing ---
beautifulsoup4>=4.12.0     # Scraping logic
lxml>=5.0.0               # Fast parser backend

# --- Data Validation & retries ---
pydantic>=2.5.0            # Schema validation (v2)
tenacity>=8.2.0            # Robust retry logic

# --- Data Science & Processing ---
pandas>=2.1.0              # DataFrame & CSV exports
numpy>=1.26.0              # Numerical operations (often used with pandas)

# --- Database Integration ---
psycopg2-binary>=2.9.0     # PostgreSQL driver (for synchronous modules)
asyncpg>=0.29.0            # Async PostgreSQL driver (for scheduler/live_pool)
SQLAlchemy>=2.0.0          # ORM & Schema management
python-dotenv>=1.0.0       # Manage .env configurations

# --- AI & LLM Providers ---
google-genai>=1.0.0        # Gemini API
groq>=0.4.0                # Llama-3 (Groq) API

# --- Feature Specific ---
feedparser>=6.0.0          # RSS News Radar
RapidFuzz>=3.0.0           # Modern fast fuzzy matching
thefuzz>=0.20.0            # Legacy fuzzy matching (fallback)
colorlog>=6.8.0            # Beautiful terminal logging
